+++ supabase/schemas/05_rls/05-06_api_keys_rls.sql
@@ -19,7 +19,7 @@ using (
     )
   )
   or organization_id in (
-    select public.get_authorized_orgs('owner')
+    select rls.get_authorized_orgs('owner')
   )
 );
 
@@ -29,7 +29,7 @@ for insert
 to authenticated, anon
 with check (
   organization_id in (
-    select public.get_authorized_orgs('owner')
+    select rls.get_authorized_orgs('owner')
   )
 );
 
@@ -39,6 +39,6 @@ for delete
 to authenticated, anon
 using (
   organization_id in (
-    select public.get_authorized_orgs('owner')
+    select rls.get_authorized_orgs('owner')
   )
 );